+++ R/DR_coxph.R
@@ -43,8 +43,8 @@
 #' @param \dots Arguments to be passed on to \code{survival::coxph}.
 #' @return \item{Named num}{Vector of the residual values.}
 #' @author Frédéric Bertrand\cr
-#' \email{frederic.bertrand@@utt.fr}\cr
-#' \url{http://www-irma.u-strasbg.fr/~fbertran/}
+#' \email{frederic.bertrand@@lecnam.net}\cr
+#' \url{https://fbertran.github.io/homepage/}
 #' @seealso \code{\link[survival]{coxph}}
 #' @references plsRcox, Cox-Models in a high dimensional setting in R, Frederic
 #' Bertrand, Philippe Bastien, Nicolas Meyer and Myriam Maumy-Bertrand (2014).