+++ R/coxsplsDR.R
@@ -80,8 +80,8 @@
 #' \code{allres=TRUE} : \item{tt_splsDR}{sPLSR components.}
 #' \item{cox_splsDR}{Final Cox-model.} \item{splsDR_mod}{The sPLSR model.}
 #' @author Frédéric Bertrand\cr
-#' \email{frederic.bertrand@@utt.fr}\cr
-#' \url{http://www-irma.u-strasbg.fr/~fbertran/}
+#' \email{frederic.bertrand@@lecnam.net}\cr
+#' \url{https://fbertran.github.io/homepage/}
 #' @seealso \code{\link[survival]{coxph}}, \code{\link[pls]{plsr}}
 #' @references plsRcox, Cox-Models in a high dimensional setting in R, Frederic
 #' Bertrand, Philippe Bastien, Nicolas Meyer and Myriam Maumy-Bertrand (2014).