+++ R/cv.coxDKplsDR.R
@@ -129,8 +129,8 @@
 #' \item{All_indics}{All results of the functions that perform error
 #' computation, for each fold, each component and error criterion.}
 #' @author Frédéric Bertrand\cr
-#' \email{frederic.bertrand@@utt.fr}\cr
-#' \url{http://www-irma.u-strasbg.fr/~fbertran/}
+#' \email{frederic.bertrand@@lecnam.net}\cr
+#' \url{https://fbertran.github.io/homepage/}
 #' @seealso See Also \code{\link{coxDKplsDR}}
 #' @references plsRcox, Cox-Models in a high dimensional setting in R, Frederic
 #' Bertrand, Philippe Bastien, Nicolas Meyer and Myriam Maumy-Bertrand (2014).