+++ R/internal-plsRcox.R
@@ -7,8 +7,8 @@
 #' @aliases getIndic getIndicCV getIndicCViAUCSH getIndicCViAUCSurvROCTest
 #' correctp.cox spls.cox ust spls.dv pls.cox predict.pls.cox
 #' @author Frédéric Bertrand\cr
-#' \email{frederic.bertrand@@utt.fr}\cr
-#' \url{http://www-irma.u-strasbg.fr/~fbertran/}
+#' \email{frederic.bertrand@@lecnam.net}\cr
+#' \url{https://fbertran.github.io/homepage/}
 #' @references plsRcox, Cox-Models in a high dimensional setting in R, Frederic
 #' Bertrand, Philippe Bastien, Nicolas Meyer and Myriam Maumy-Bertrand (2014).
 #' Proceedings of User2014!, Los Angeles, page 152.\cr